requirements.txt
distro==1.9.0
psutil==6.0.0
py-cpuinfo==9.0.0
pystyle==2.9
